wutdevoptab: Rename all files to *fsa*.cpp
