+++ libraries/wutdevoptab/devoptab_fsa_dirreset.cpp
@@ -1,4 +1,4 @@
-#include "devoptab_fs.h"
+#include "devoptab_fsa.h"
 
 int
 __wut_fs_dirreset(struct _reent *r,
