+++ libraries/wutdevoptab/devoptab_fsa_rename.cpp
@@ -1,4 +1,4 @@
-#include "devoptab_fs.h"
+#include "devoptab_fsa.h"
 
 int
 __wut_fs_rename(struct _reent *r,
