+++ libraries/wutdevoptab/devoptab_fsa_unlink.cpp
@@ -1,4 +1,4 @@
-#include "devoptab_fs.h"
+#include "devoptab_fsa.h"
 
 int
 __wut_fs_unlink(struct _reent *r,
